requirements.txt
# MCP framework
mcp[cli]>=1.0.0
openai>=1.0.0

# Deep learning
torch>=2.0.0
transformers>=4.40.0
fair-esm>=2.0.0

# Data
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0

# Utilities
tqdm>=4.65.0
scikit-learn>=1.3.0
